extract interface FutureResponseContextStore
reason: to enable multiple implementations
